requirements.txt
# Agentic Clinical Assistant - Production Dependencies
# Generated from pyproject.toml
# Python >=3.10 required

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Vector Databases
faiss-cpu>=1.7.4
pinecone-client>=2.2.4
weaviate-client>=3.25.0

# LLM SDKs
openai>=1.3.0
anthropic>=0.7.0

# Embeddings
sentence-transformers>=2.2.0

# Monitoring & Observability
prometheus-client>=0.19.0
structlog>=23.2.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
tenacity>=8.2.0
python-multipart>=0.0.6

